LUCENE-9302: Grouping to use long to avoid overflows

+++ lucene/grouping/src/test/org/apache/lucene/search/grouping/TestGrouping.java
@@ -504,7 +504,7 @@ public class TestGrouping extends LuceneTestCase {
     final List<BytesRef> sortedGroups = new ArrayList<>();
     final List<Comparable<?>[]> sortedGroupFields = new ArrayList<>();
 
-    int totalHitCount = 0;
+    long totalHitCount = 0;
     Set<BytesRef> knownGroups = new HashSet<>();
 
     // System.out.println("TEST: slowGrouping");
@@ -544,7 +544,7 @@ public class TestGrouping extends LuceneTestCase {
     final Comparator<GroupDoc> docSortComp = getComparator(docSort);
     @SuppressWarnings({"unchecked", "rawtypes"})
     final GroupDocs<BytesRef>[] result = new GroupDocs[limit - groupOffset];
-    int totalGroupedHitCount = 0;
+    long totalGroupedHitCount = 0;
     for (int idx = groupOffset; idx < limit; idx++) {
       final BytesRef group = sortedGroups.get(idx);
       final List<GroupDoc> docs = groups.get(group);
@@ -583,7 +583,7 @@ public class TestGrouping extends LuceneTestCase {
               totalGroupedHitCount,
               result,
               Float.NaN),
-          knownGroups.size());
+          (long) knownGroups.size());
     } else {
       return new TopGroups<>(
           groupSort.getSort(),
@@ -1093,7 +1093,8 @@ public class TestGrouping extends LuceneTestCase {
 
           if (doAllGroups) {
             TopGroups<BytesRef> tempTopGroups = getTopGroups(c2, docOffset);
-            groupsResult = new TopGroups<>(tempTopGroups, allGroupsCollector.getGroupCount());
+            groupsResult =
+                new TopGroups<>(tempTopGroups, (long) allGroupsCollector.getGroupCount());
           } else {
             groupsResult = getTopGroups(c2, docOffset);
           }
@@ -1218,9 +1219,10 @@ public class TestGrouping extends LuceneTestCase {
         final TopGroups<BytesRef> groupsResultBlocks;
         if (doAllGroups && tempTopGroupsBlocks != null) {
           assertEquals(
-              (int) tempTopGroupsBlocks.totalGroupCount, allGroupsCollector2.getGroupCount());
+              (long) tempTopGroupsBlocks.totalGroupCount,
+              (long) allGroupsCollector2.getGroupCount());
           groupsResultBlocks =
-              new TopGroups<>(tempTopGroupsBlocks, allGroupsCollector2.getGroupCount());
+              new TopGroups<>(tempTopGroupsBlocks, (long) allGroupsCollector2.getGroupCount());
         } else {
           groupsResultBlocks = tempTopGroupsBlocks;
         }
